Util Updates
We aren't using AdvantageKit for subsystems that use the ConstantsIO types. There are a lot of issues with how those data types work. Happy to have a more in depth conversation on this.

Replaced the ConstantsIO types with immutable data types using the Fluent Builder Pattern. Please read up on this pattern. There are examples of how these are constructed in Constants and how they are used in Elevator and Shooter.

I also added a bunch of TODOs in the Shooter and Elevator.